[REF] move estimator checks in their own module
