+++ nilearn/_utils/tests/test_class_inspect.py
@@ -6,7 +6,6 @@ which we historically used,
 ignores modules whose name starts with an underscore.
 """
 
-import pytest
 from sklearn.base import BaseEstimator
 
 from nilearn._utils import class_inspect
@@ -36,31 +35,3 @@ def test_get_params():
     assert params_a_in_b == {"a": 1}
     params_a_in_b = class_inspect.get_params(A, b, ignore=["a"])
     assert params_a_in_b == {}
-
-
-def test_check_estimator_has_sklearn_is_fitted():
-    """Check errors are thrown for unfitted estimator.
-
-    Check that before fitting
-    - estimlator has a __sklearn_is_fitted__ method that returns false
-    - running sklearn check_is_fitted on masker throws an error
-    """
-
-    class DummyEstimator:
-        def __init__(self):
-            pass
-
-    with pytest.raises(
-        TypeError, match="must have __sklearn_is_fitted__ method"
-    ):
-        class_inspect.check_estimator_has_sklearn_is_fitted(DummyEstimator())
-
-    class DummyEstimator:
-        def __init__(self):
-            pass
-
-        def __sklearn_is_fitted__(self):
-            return True
-
-    with pytest.raises(ValueError, match="must return False before fit"):
-        class_inspect.check_estimator_has_sklearn_is_fitted(DummyEstimator())
