+++ nilearn/connectome/tests/test_group_sparse_cov.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from nilearn._utils.class_inspect import check_estimator
 from nilearn._utils.data_gen import generate_group_sparse_gaussian_graphs
+from nilearn._utils.estimator_checks import check_estimator
 from nilearn.connectome import GroupSparseCovariance, GroupSparseCovarianceCV
 from nilearn.connectome.group_sparse_cov import (
     group_sparse_covariance,
