+++ nilearn/maskers/tests/test_nifti_spheres_masker.py
@@ -8,7 +8,7 @@ from nibabel import Nifti1Image
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 
 from nilearn._utils import data_gen
-from nilearn._utils.class_inspect import check_estimator
+from nilearn._utils.estimator_checks import check_estimator
 from nilearn.image import get_data, new_img_like
 from nilearn.maskers import NiftiSpheresMasker
