+++ nilearn/maskers/tests/test_surface_masker.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from nilearn._utils.class_inspect import check_estimator
+from nilearn._utils.estimator_checks import check_estimator
 from nilearn.maskers import SurfaceMasker
 from nilearn.surface import SurfaceImage
 from nilearn.surface._testing import (
