+++ nilearn/regions/tests/test_region_extractor.py
@@ -5,8 +5,8 @@ import pytest
 from nibabel import Nifti1Image
 from scipy.ndimage import label
 
-from nilearn._utils.class_inspect import check_estimator
 from nilearn._utils.data_gen import generate_labeled_regions, generate_maps
+from nilearn._utils.estimator_checks import check_estimator
 from nilearn._utils.exceptions import DimensionError
 from nilearn.conftest import _img_4d_zeros
 from nilearn.image import get_data
